requirements.txt
fastapi==0.128.0
httpx==0.28.1
python-dotenv==1.2.1
Requests==2.32.5
SQLAlchemy==2.0.45
sqlmodel==0.0.31
streamlit==1.52.2
uvicorn==0.38.0
pymysql==1.1.2
cryptography==46.0.3